Move SSM config fetch to its own module
